changed id from int to long

+++ src/main/java/org/atlasapi/media/channel/ChannelGroupWriter.java
@@ -4,5 +4,5 @@ public interface ChannelGroupWriter {
     
     ChannelGroup createOrUpdate(ChannelGroup channelGroup);
 
-    void deleteChannelGroupById(int channelGroupId);
+    void deleteChannelGroupById(long channelGroupId);
 }
